Разделил тесты по разным классам - каждый под свою ручку
